# Misc/CastleSwaggerApi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Demo API",
    "version": "v1"
  },
  "paths": {
...
    "/soldier/{soldierId}": {
      "get": {
        "tags": [
          "Knight"
        ],
        "operationId": "GetKnight",
        "parameters": [
          {
            "name": "soldierId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Knight"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Knight"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Knight"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Knight"
        ],
        "operationId": "TaskKnight",
        "parameters": [
          {
            "name": "soldierId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskKnightRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskKnightRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TaskKnightRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TaskMessage"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskMessage"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskMessage"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Knight"
        ],
        "operationId": "RemoveKnight",
        "parameters": [
          {
            "name": "soldierId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/soldier/{soldierId}/tasks": {
      "get": {
        "tags": [
          "Knight"
        ],
        "operationId": "GetTaskResults",
        "parameters": [
          {
            "name": "soldierId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskResultMessage"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskResultMessage"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskResultMessage"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/soldier/{soldierId}/tasks/{taskId}": {
      "get": {
        "tags": [
          "Knight"
        ],
        "operationId": "GetTaskResult",
        "parameters": [
          {
            "name": "soldierId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResultMessage"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResultMessage"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResultMessage"
                }
              }
            }
          }
        }
      }
    },
    "/soldier/{soldierId}/historys": {
      "get": {
        "tags": [
          "Knight"
        ],
        "operationId": "GetKnightHistorys",
        "parameters": [
          {
            "name": "soldierId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KnightHistory"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KnightHistory"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KnightHistory"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/soldier/{soldierId}/history/{taskId}": {
      "get": {
        "tags": [
          "Knight"
        ],
        "operationId": "GetKnightHistory",
        "parameters": [
          {
            "name": "soldierId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/KnightHistory"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnightHistory"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnightHistory"
                }
              }
            }
          }
        }
      }
    },
    "/soldier/{soldierId}/historyresult/{taskId}": {
      "get": {
        "tags": [
          "Knight"
        ],
        "operationId": "GetHistoryResult",
        "parameters": [
          {
            "name": "soldierId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResultMessage"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResultMessage"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResultMessage"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AllyDirection": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "Integrity": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "KNIGHT_TASK_STATUS": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "Knight": {
        "type": "object",
        "properties": {
          "metadata": {
            "$ref": "#/components/schemas/KnightMetadata"
          },
          "firstSeen": {
            "type": "string",
            "format": "date-time"
          },
          "lastSeen": {
            "type": "string",
            "format": "date-time"
          },
          "parentId": {
            "type": "string",
            "nullable": true
          },
          "direction": {
            "$ref": "#/components/schemas/AllyDirection"
          }
        },
        "additionalProperties": false
      },
      "KnightHistory": {
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string",
            "nullable": true
          },
          "knightId": {
            "type": "string",
            "nullable": true
          },
          "command": {
            "type": "string",
            "nullable": true
          },
          "arguments": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "file": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "taskSentAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/KNIGHT_TASK_STATUS"
          },
          "result": {
            "type": "string",
            "nullable": true
          },
          "resultsAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "KnightMetadata": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "hostname": {
            "type": "string",
            "nullable": true
          },
          "integrity": {
            "$ref": "#/components/schemas/Integrity"
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "processName": {
            "type": "string",
            "nullable": true
          },
          "processId": {
            "type": "integer",
            "format": "int32"
          },
          "x64": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "TaskKnightRequest": {
        "type": "object",
        "properties": {
          "command": {
            "type": "string",
            "nullable": true
          },
          "arguments": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "file": {
            "type": "string",
            "format": "byte",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TaskMessage": {
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string",
            "nullable": true
          },
          "command": {
            "type": "string",
            "nullable": true
          },
          "arguments": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "file": {
            "type": "string",
            "format": "byte",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TaskResultMessage": {
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string",
            "nullable": true
          },
          "result": {
            "type": "string",
            "nullable": true
          },
          "completetedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Please enter a valid token",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}